+++ app/src/main/java/com/boardgamegeek/entities/HIndexEntity.kt
@@ -5,9 +5,11 @@ data class HIndexEntity(val h: Int, val n: Int) {
         get() = h + 1 - n.toDouble() / (2 * h + 1)
 
     val description: String
-        get() = String.format("%.2f", rational)
+        get() = if (h == INVALID_H_INDEX) "?" else String.format("%.2f", rational)
 
     companion object {
+        const val INVALID_H_INDEX = -1
+
         fun fromList(list: List<Int>): HIndexEntity {
             val counts = list.filter { it > 0 }.sortedDescending()