+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 6.7
  * Lots more data for designers, artists, publishers, mechanics, and categories
  * UI changes
  * View and edit inventory location
+ * H-index includes rational value indicating how close you are to the next number
  * Improve search usability
  * Ability to specify negative scores more easily
  * Show collection status of expansion and base games